# Conflicts:
#	src/Extensions/RepeatedTest.php
#	src/TextUI/Command.php

+++ src/TextUI/Command.php
@@ -12,13 +12,7 @@
  * A TestRunner for the Command Line Interface (CLI)
  * PHP SAPI Module.
  *
- * @package    PHPUnit
- * @subpackage TextUI
- * @author     Sebastian Bergmann <sebastian@phpunit.de>
- * @copyright  Sebastian Bergmann <sebastian@phpunit.de>
- * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
- * @link       http://www.phpunit.de/
- * @since      Class available since Release 3.0.0
+ * @since Class available since Release 3.0.0
  */
 class PHPUnit_TextUI_Command
 {
@@ -26,9 +20,9 @@ class PHPUnit_TextUI_Command
      * @var array
      */
     protected $arguments = array(
-      'listGroups'              => false,
-      'loader'                  => null,
-      'useDefaultConfiguration' => true
+        'listGroups'              => false,
+        'loader'                  => null,
+        'useDefaultConfiguration' => true
     );
 
     /**
@@ -40,64 +34,65 @@ class PHPUnit_TextUI_Command
      * @var array
      */
     protected $longOptions = array(
-      'colors==' => null,
-      'bootstrap=' => null,
-      'columns=' => null,
-      'configuration=' => null,
-      'coverage-clover=' => null,
-      'coverage-crap4j=' => null,
-      'coverage-html=' => null,
-      'coverage-php=' => null,
-      'coverage-text==' => null,
-      'coverage-xml=' => null,
-      'debug' => null,
-      'exclude-group=' => null,
-      'filter=' => null,
-      'testsuite=' => null,
-      'group=' => null,
-      'help' => null,
-      'include-path=' => null,
-      'list-groups' => null,
-      'loader=' => null,
-      'log-json=' => null,
-      'log-junit=' => null,
-      'log-tap=' => null,
-      'process-isolation' => null,
-      'repeat=' => null,
-      'only-repeat-failed' => null,
-      'stderr' => null,
-      'stop-on-error' => null,
-      'stop-on-failure' => null,
-      'stop-on-incomplete' => null,
-      'stop-on-risky' => null,
-      'stop-on-skipped' => null,
-      'report-useless-tests' => null,
-      'strict-coverage' => null,
-      'disallow-test-output' => null,
-      'enforce-time-limit' => null,
-      'disallow-todo-tests' => null,
-      'strict-global-state' => null,
-      'strict' => null,
-      'tap' => null,
-      'testdox' => null,
-      'testdox-html=' => null,
-      'testdox-text=' => null,
-      'test-suffix=' => null,
-      'no-configuration' => null,
-      'no-globals-backup' => null,
-      'printer=' => null,
-      'static-backup' => null,
-      'verbose' => null,
-      'version' => null
+        'colors=='             => null,
+        'bootstrap='           => null,
+        'columns='             => null,
+        'configuration='       => null,
+        'coverage-clover='     => null,
+        'coverage-crap4j='     => null,
+        'coverage-html='       => null,
+        'coverage-php='        => null,
+        'coverage-text=='      => null,
+        'coverage-xml='        => null,
+        'debug'                => null,
+        'exclude-group='       => null,
+        'filter='              => null,
+        'testsuite='           => null,
+        'group='               => null,
+        'help'                 => null,
+        'include-path='        => null,
+        'list-groups'          => null,
+        'loader='              => null,
+        'log-json='            => null,
+        'log-junit='           => null,
+        'log-tap='             => null,
+        'process-isolation'    => null,
+        'repeat='              => null,
+        'only-repeat-failed'   => null,
+        'stderr'               => null,
+        'stop-on-error'        => null,
+        'stop-on-failure'      => null,
+        'stop-on-incomplete'   => null,
+        'stop-on-risky'        => null,
+        'stop-on-skipped'      => null,
+        'report-useless-tests' => null,
+        'strict-coverage'      => null,
+        'disallow-test-output' => null,
+        'enforce-time-limit'   => null,
+        'disallow-todo-tests'  => null,
+        'strict-global-state'  => null,
+        'strict'               => null,
+        'tap'                  => null,
+        'testdox'              => null,
+        'testdox-html='        => null,
+        'testdox-text='        => null,
+        'test-suffix='         => null,
+        'no-configuration'     => null,
+        'no-coverage'          => null,
+        'no-globals-backup'    => null,
+        'printer='             => null,
+        'static-backup'        => null,
+        'verbose'              => null,
+        'version'              => null
     );
 
     /**
-     * @var boolean
+     * @var bool
      */
     private $versionStringPrinted = false;
 
     /**
-     * @param boolean $exit
+     * @param bool $exit
      */
     public static function main($exit = true)
     {
@@ -107,9 +102,10 @@ class PHPUnit_TextUI_Command
     }
 
     /**
-     * @param  array $argv
-     * @param  boolean $exit
-     * @return integer
+     * @param array $argv
+     * @param bool  $exit
+     *
+     * @return int
      */
     public function run(array $argv, $exit = true)
     {
@@ -175,6 +171,7 @@ class PHPUnit_TextUI_Command
      * Create a TestRunner, override in subclasses.
      *
      * @return PHPUnit_TextUI_TestRunner
+     *
      * @since  Method available since Release 3.6.0
      */
     protected function createRunner()
@@ -230,8 +227,11 @@ class PHPUnit_TextUI_Command
     protected function handleArguments(array $argv)
     {
         if (defined('__PHPUNIT_PHAR__')) {
-            $this->longOptions['selfupdate']  = null;
-            $this->longOptions['self-update'] = null;
+            $this->longOptions['check-version'] = null;
+            $this->longOptions['selfupdate']    = null;
+            $this->longOptions['self-update']   = null;
+            $this->longOptions['selfupgrade']   = null;
+            $this->longOptions['self-upgrade']  = null;
         }
 
         try {
@@ -288,9 +288,9 @@ class PHPUnit_TextUI_Command
                         $option[1] = 'php://stdout';
                     }
 
-                    $this->arguments['coverageText'] = $option[1];
+                    $this->arguments['coverageText']                   = $option[1];
                     $this->arguments['coverageTextShowUncoveredFiles'] = false;
-                    $this->arguments['coverageTextShowOnlySummary'] = false;
+                    $this->arguments['coverageTextShowOnlySummary']    = false;
                     break;
 
                 case '--coverage-xml':
@@ -317,7 +317,7 @@ class PHPUnit_TextUI_Command
                 case '--help':
                     $this->showHelp();
                     exit(PHPUnit_TextUI_TestRunner::SUCCESS_EXIT);
-                break;
+                    break;
 
                 case '--filter':
                     $this->arguments['filter'] = $option[1];
@@ -429,6 +429,10 @@ class PHPUnit_TextUI_Command
                     $this->arguments['useDefaultConfiguration'] = false;
                     break;
 
+                case '--no-coverage':
+                    $this->arguments['noCoverage'] = true;
+                    break;
+
                 case '--no-globals-backup':
                     $this->arguments['backupGlobals'] = false;
                     break;
@@ -445,7 +449,7 @@ class PHPUnit_TextUI_Command
                 case '--version':
                     $this->printVersionString();
                     exit(PHPUnit_TextUI_TestRunner::SUCCESS_EXIT);
-                break;
+                    break;
 
                 case '--report-useless-tests':
                     $this->arguments['reportUselessTests'] = true;
@@ -480,9 +484,22 @@ class PHPUnit_TextUI_Command
                     $this->arguments['deprecatedStrictModeOption'] = true;
                     break;
 
+                case '--check-version':
+                    $this->handleVersionCheck();
+                    break;
+
                 case '--selfupdate':
                 case '--self-update':
                     $this->handleSelfUpdate();
+                    break;
+
+                case '--selfupgrade':
+                case '--self-upgrade':
+                    $this->handleSelfUpdate(true);
+                    break;
+
+                case '--whitelist':
+                    $this->arguments['whitelist'] = $option[1];
                     break;
 
                 default:
@@ -538,8 +555,7 @@ class PHPUnit_TextUI_Command
 
         if (isset($this->arguments['configuration']) &&
             is_dir($this->arguments['configuration'])) {
-            $configurationFile = $this->arguments['configuration'] .
-                                 '/phpunit.xml';
+            $configurationFile = $this->arguments['configuration'] . '/phpunit.xml';
 
             if (file_exists($configurationFile)) {
                 $this->arguments['configuration'] = realpath(
@@ -551,7 +567,7 @@ class PHPUnit_TextUI_Command
                 );
             }
         } elseif (!isset($this->arguments['configuration']) &&
-                 $this->arguments['useDefaultConfiguration']) {
+                  $this->arguments['useDefaultConfiguration']) {
             if (file_exists('phpunit.xml')) {
                 $this->arguments['configuration'] = realpath('phpunit.xml');
             } elseif (file_exists('phpunit.xml.dist')) {
@@ -566,6 +582,9 @@ class PHPUnit_TextUI_Command
                 $configuration = PHPUnit_Util_Configuration::getInstance(
                     $this->arguments['configuration']
                 );
+            } catch (Throwable $e) {
+                print $e->getMessage() . "\n";
+                exit(PHPUnit_TextUI_TestRunner::FAILURE_EXIT);
             } catch (Exception $e) {
                 print $e->getMessage() . "\n";
                 exit(PHPUnit_TextUI_TestRunner::FAILURE_EXIT);
@@ -575,7 +594,7 @@ class PHPUnit_TextUI_Command
 
             $configuration->handlePHPConfiguration();
 
-            /**
+            /*
              * Issue #1216
              */
             if (isset($this->arguments['bootstrap'])) {
@@ -584,11 +603,15 @@ class PHPUnit_TextUI_Command
                 $this->handleBootstrap($phpunit['bootstrap']);
             }
 
-            /**
+            /*
              * Issue #657
              */
             if (isset($phpunit['stderr']) && ! isset($this->arguments['stderr'])) {
                 $this->arguments['stderr'] = $phpunit['stderr'];
+            }
+
+            if (isset($phpunit['columns']) && ! isset($this->arguments['columns'])) {
+                $this->arguments['columns'] = $phpunit['columns'];
             }
 
             if (isset($phpunit['printerClass'])) {
@@ -619,9 +642,12 @@ class PHPUnit_TextUI_Command
 
             $browsers = $configuration->getSeleniumBrowserConfiguration();
 
-            if (!empty($browsers) &&
-                class_exists('PHPUnit_Extensions_SeleniumTestCase')) {
-                PHPUnit_Extensions_SeleniumTestCase::$browsers = $browsers;
+            if (!empty($browsers)) {
+                $this->arguments['deprecatedSeleniumConfiguration'] = true;
+
+                if (class_exists('PHPUnit_Extensions_SeleniumTestCase')) {
+                    PHPUnit_Extensions_SeleniumTestCase::$browsers = $browsers;
+                }
             }
 
             if (!isset($this->arguments['test'])) {
@@ -657,8 +683,9 @@ class PHPUnit_TextUI_Command
     /**
      * Handles the loading of the PHPUnit_Runner_TestSuiteLoader implementation.
      *
-     * @param  string                         $loaderClass
-     * @param  string                         $loaderFile
+     * @param string $loaderClass
+     * @param string $loaderFile
+     *
      * @return PHPUnit_Runner_TestSuiteLoader
      */
     protected function handleLoader($loaderClass, $loaderFile = '')
@@ -701,8 +728,9 @@ class PHPUnit_TextUI_Command
     /**
      * Handles the loading of the PHPUnit_Util_Printer implementation.
      *
-     * @param  string               $printerClass
-     * @param  string               $printerFile
+     * @param string $printerClass
+     * @param string $printerFile
+     *
      * @return PHPUnit_Util_Printer
      */
     protected function handlePrinter($printerClass, $printerFile = '')
@@ -762,22 +790,40 @@ class PHPUnit_TextUI_Command
     /**
      * @since Method available since Release 4.0.0
      */
-    protected function handleSelfUpdate()
+    protected function handleSelfUpdate($upgrade = false)
     {
         $this->printVersionString();
+
+        $localFilename = realpath($_SERVER['argv'][0]);
+
+        if (!is_writable($localFilename)) {
+            print 'No write permission to update ' . $localFilename . "\n";
+            exit(PHPUnit_TextUI_TestRunner::EXCEPTION_EXIT);
+        }
 
         if (!extension_loaded('openssl')) {
             print "The OpenSSL extension is not loaded.\n";
             exit(PHPUnit_TextUI_TestRunner::EXCEPTION_EXIT);
         }
 
-        $remoteFilename = sprintf(
-            'https://phar.phpunit.de/phpunit%s.phar',
-            PHPUnit_Runner_Version::getReleaseChannel()
-        );
+        if (!$upgrade) {
+            $remoteFilename = sprintf(
+                'https://phar.phpunit.de/phpunit-%s.phar',
+                file_get_contents(
+                    sprintf(
+                        'https://phar.phpunit.de/latest-version-of/phpunit-%s',
+                        PHPUnit_Runner_Version::series()
+                    )
+                )
+            );
+        } else {
+            $remoteFilename = sprintf(
+                'https://phar.phpunit.de/phpunit%s.phar',
+                PHPUnit_Runner_Version::getReleaseChannel()
+            );
+        }
 
-        $localFilename = realpath($_SERVER['argv'][0]);
-        $tempFilename  = basename($localFilename, '.phar') . '-temp.phar';
+        $tempFilename = tempnam(sys_get_temp_dir(), 'phpunit') . '.phar';
 
         // Workaround for https://bugs.php.net/bug.php?id=65538
         $caFile = dirname($tempFilename) . '/ca.pem';
@@ -788,8 +834,8 @@ class PHPUnit_TextUI_Command
         $options = array(
             'ssl' => array(
                 'allow_self_signed' => false,
-                'cafile' => $caFile,
-                'verify_peer' => true
+                'cafile'            => $caFile,
+                'verify_peer'       => true
             )
         );
 
@@ -814,7 +860,13 @@ class PHPUnit_TextUI_Command
             unset($phar);
             rename($tempFilename, $localFilename);
             unlink($caFile);
-        } catch (Exception $e) {
+        } catch (Throwable $_e) {
+            $e = $_e;
+        } catch (Exception $_e) {
+            $e = $_e;
+        }
+
+        if (isset($e)) {
             unlink($caFile);
             unlink($tempFilename);
             print " done\n\n" . $e->getMessage() . "\n";
@@ -822,7 +874,27 @@ class PHPUnit_TextUI_Command
         }
 
         print " done\n";
-        exit(0);
+        exit(PHPUnit_TextUI_TestRunner::SUCCESS_EXIT);
+    }
+
+    /**
+     * @since Method available since Release 4.8.0
+     */
+    protected function handleVersionCheck()
+    {
+        $this->printVersionString();
+
+        $latestVersion = file_get_contents('https://phar.phpunit.de/latest-version-of/phpunit');
+        $isOutdated    = version_compare($latestVersion, PHPUnit_Runner_Version::id(), '>');
+
+        if ($isOutdated) {
+            print "You are not using the latest version of PHPUnit.\n";
+            print 'Use "phpunit --self-upgrade" to install PHPUnit ' . $latestVersion . "\n";
+        } else {
+            print "You are using the latest version of PHPUnit.\n";
+        }
+
+        exit(PHPUnit_TextUI_TestRunner::SUCCESS_EXIT);
     }
 
     /**
@@ -901,6 +973,7 @@ Configuration Options:
   --bootstrap <file>        A "bootstrap" PHP file that is run before the tests.
   -c|--configuration <file> Read configuration from XML file.
   --no-configuration        Ignore default configuration file (phpunit.xml).
+  --no-coverage             Ignore code coverage configuration.
   --include-path <path(s)>  Prepend PHP's include_path with given path(s).
   -d key[=value]            Sets a php.ini value.
 
@@ -912,7 +985,9 @@ Miscellaneous Options:
 EOT;
 
         if (defined('__PHPUNIT_PHAR__')) {
-            print "\n  --self-update             Update PHPUnit to the latest version.\n";
+            print "\n  --check-version           Check whether PHPUnit is the latest version.";
+            print "\n  --self-update             Update PHPUnit to the latest version within the same\n                            release series.\n";
+            print "\n  --self-upgrade            Upgrade PHPUnit to the latest version.\n";
         }
     }
